Enable ruff's default rules and update ruff to 0.16.1
lint.select replaced ruff's own defaults with just the eight listed prefixes,
so whole families it enables out of the box were silently off. Switching to
lint.extend-select keeps the defaults and adds ours on top, taking the enabled
set from 205 rules to 467 and surfacing 161 findings.

Fixed rather than ignored, everything that was a real improvement:

  ISC004  implicit string concatenation inside a collection, where a missing
          comma would silently merge two elements. The chat_threads code
          templates and one timescaledb docker argument are now explicitly
          parenthesised; template contents verified byte-identical.
  TRY401  logging.exception already appends the traceback, whose last line is
          the exception type and str(exc), so interpolating {exc} printed the
          same text twice. The sibling reason= strings keep it, as those go to
          the results database where there is no traceback.
  RUF100  two noqa directives for a rule that is not enabled
  RUF046  int() around a round() that already returns an int
  S112    a swallowed reflection failure in alembic/env.py now logs at debug
  RET501  redundant return None
  FURB157 Decimal(str) where an int literal is exact

Two families are ignored because they conflict with deliberate design, with the
reasoning recorded in pyproject.toml:

  DTZ     timestamps are stored naive throughout, following the DuckDB TIMESTAMP
          and datetime.now(UTC).replace(tzinfo=None) convention
  BLE001  readiness polling, per-query suite error handling and the sampling
          lanes all have to survive arbitrary driver failures

RUF012 is a false positive for the test doubles: they subclass the pydantic
Database, where a mutable default is a field default that pydantic copies per
instance, so ClassVar would remove them as fields and break the fakes. Verified
by construction and per-file-ignored. The two genuine cases, a plain formatter
class and FakeStorage, are annotated ClassVar.

pre-commit hooks move to ruff-pre-commit v0.16.1 and pre-commit-hooks v6.0.0.

+++ olap_benchmarks/suites/chat_threads/generator.py
@@ -445,7 +445,7 @@ def _thread_message_count(rng: Random, is_anchor: bool) -> int:
     if is_anchor:
         return rng.randrange(ANCHOR_MIN_MESSAGES, ANCHOR_MAX_MESSAGES + 1)
 
-    drawn = int(round(rng.lognormvariate(np.log(MEAN_THREAD_MESSAGES) - 0.5 * 0.8**2, 0.8)))
+    drawn = round(rng.lognormvariate(np.log(MEAN_THREAD_MESSAGES) - 0.5 * 0.8**2, 0.8))
     return max(MIN_THREAD_MESSAGES, min(MAX_THREAD_MESSAGES, drawn))
